fix watchers on elasticache clustered mode (#42)
Right now when omnibot is deployed with elasticache as its redis store configured in clustered mode the locks in [python-redis-lock](https://github.com/ionelmc/python-redis-lock) can fail because their keys can get hashed to different clusters, causing an error: 

```
redis.exceptions.ResponseError: CROSSSLOT Keys in request don't hash to the same slot
```

According to [AWS guidance on this error](https://aws.amazon.com/premiumsupport/knowledge-center/elasticache-crossslot-keys-error-redis/) you can force keys to be hashed to the same cluster by including a substring surrounded by `{braces}` such that only the text _inside_ the braces is used for hashing. This PR does exactly that on all of the Redis locks.

Note that this _is a breaking change_, and will, worst case, cause 2+ instances of a watcher to reindex users/conversations/emojis. That should be fine, though, as the data will become coherent again the next time it's reindexed.