Allow synchronous callbacks to be regsitered for interactive components
Slack (`view_submission`)[https://api.slack.com/surfaces/modals/using#handling_submissions] handlers require responses to the callback request to be able to render errors.

This change allows a given (omnibot) callback registered against an interactive component to be registered as `synchronous`, in which case the result of the callback will immediately be provided to the slack callback.

A (interactive component) callback can be set to synchronous by setting the `synchronous` key to `true` for a given callback.

```yaml
    - callback_id: "modal-with-inputs"
      response_type: "in_channel"
      bots:
        "test-team-name":
          - "TEST_OMNIBOT_NAME"
        "testteam":
          - "echobot"
      callbacks:
        - module: "tests.data:test_handler"
          synchronous: true
```

Note that as only one callback can be used as a response, any further callbacks registered with a given callback ID will be ignored.