[DEVCON-1527] Account for 1:1 direct messages in cache (#81)
We saw the following error in the logs and realized that the issue may
be because of the different responses that Slack returns.
```
Traceback (most recent call last):
  File "/code/omnibot-private/omnibot/processor.py", line 52, in process_event
    message = Message(bot, event_info, event_trace)
  File "/code/omnibot-private/omnibot/services/slack/message.py", line 54, in __init__
    self._payload['channel'] = slack.get_channel(
  File "/code/omnibot-private/omnibot/services/slack/__init__.py", line 333, in get_channel
    update_channel(bot, channel_data['channel'])
  File "/code/omnibot-private/omnibot/services/slack/__init__.py", line 136, in update_channel
    channel['name'],
KeyError: 'name'
```

The code assumes that the response for "conversations.info" has a `name`
key, but the [Slack
docs](https://api.slack.com/methods/conversations.info#examples) show
that the key only exists for public channels, private channels, and
group DMs but not for 1:1 DMs. We should update the code to account for
the last case.

See [Slack
thread](https://lyft.slack.com/archives/C017T9A506P/p1683851495598739)
for our full debugging process.