Ignore bot messages from slackbot user (#90)
On a few occasions our bot has received messages from `Slackbot`, which
results in thousands of exchanges to our bot as it's interpreting
Slakbot's message as a command request.

There's code in Omnibot to ignore messages from bots, and I originally
thought there was a bug with this logic.
https://github.com/lyft/omnibot/blob/60d042ec55295d139b92e29d737b4a4c9d1076a8/omnibot/services/slack/message.py#L58-L62

However, in inspecting Omnibot's message event to our bot, I noticed
that in the `parsed_user` object, the `is_bot` field is `false`, which
lead me to believe that Slack isn't classifying `Slackbot` as a bot user
and not sending `bot_id`. Example:
```
"parsedUser": {
            "id": "USLACKBOT",
            "tz": "America/Los_Angeles",
            "name": "slackbot",
            "color": "757575",
            "is_bot": false,
            "deleted": false,
            "profile": {
                   ..
            },
         ...
        },
```

This is confirmed [in docs](https://api.slack.com/types/user)(search for
`is_bot`) and in this
[tweet](https://twitter.com/slackapi/status/938378388871499776?lang=gl)
from Slack. PR adds a change to ignore messages from Slackbot, whose
user id will always be `USLACKBOT`.