[INTPROD-9204] Add Auth-Test Call (#107)
This pull request includes a significant change to the
`omnibot/services/slack/__init__.py` file. A new function
`get_auth(bot)` has been added. This function retrieves the
authentication information for the bot. If the authentication
information is not found in the Redis client, it makes an API call to
`auth.test`. If the API call is successful, the authentication
information is stored in the Redis client for future use. If the API
call fails, a warning is logged and an empty dictionary is returned.

*
[`omnibot/services/slack/__init__.py`](diffhunk://#diff-54762f953f73c007a82df8f886ed4ae4e222223ed69f4f5b7b7988e38de5becbR476-R499):
Added a new function `get_auth(bot)`. This function retrieves the bot's
authentication information from the Redis client, or makes an API call
to `auth.test` if the information is not found. The result of the API
call is then stored in the Redis client. If the API call fails, a
warning is logged and an empty dictionary is returned.